[Fix #155900201] fix tests
