+++ src/__tests__/Login.test.js
@@ -2,17 +2,14 @@ import React from 'react';
 import { shallow } from 'enzyme';
 import { shallowToJson } from 'enzyme-to-json';
 
-import Register from '../users/Register';
+import Login from '../components/users/Login';
 
-describe('Register component', () => {
-  const wrapper = shallow(<Register />);
+describe('Login component', () => {
+  const wrapper = shallow(<Login />);
   const preventDefault = jest.fn();
 
   it('renders properly', () => {
     expect(shallowToJson(wrapper)).toMatchSnapshot();
+    expect(wrapper.instance().loginHandler({ preventDefault }));
   });
-
-  it('has divs', () => {
-    expect(wrapper.find('div')).toHaveLength(18);
-  })
 })
